Moved CSV header definitions to validations
