+++ src/lib/validations/add-students.ts
@@ -1,0 +1,1 @@
+export const addStudentsCsvHeaders = ["full_name", "university_id", "email"];
