+++ src/lib/validations/add-supervisors.ts
@@ -1,0 +1,7 @@
+export const addSupervisorsCsvHeaders = [
+  "full_name",
+  "university_id",
+  "email",
+  "project_target",
+  "project_upper_quota",
+];
